requirements.txt
beepy==1.0.7
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
comtypes==1.2.0
fuzzywuzzy==0.18.0
google-ai-generativelanguage==0.2.0
google-api-core==2.11.1
google-auth==2.22.0
google-generativeai==0.1.0
googleapis-common-protos==1.60.0
grpcio==1.56.2
grpcio-status==1.56.2
idna==3.4
proto-plus==1.22.3
protobuf==4.24.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pypiwin32==223
python-dotenv==1.0.0
pyttsx3==2.90
pywin32==306
requests==2.31.0
rsa==4.9
simpleaudio==1.0.4
six==1.16.0
SpeechRecognition==3.10.0
urllib3==1.26.16
